Initial conversion to new graphics api
