+++ rte/_build.sh
@@ -2,4 +2,5 @@ git submodule update --init --recursive
 mkdir build
 cd build
 cmake .. -DGENERATE_WRAPPER=ON
-make
+cd ..
+cmake --build build
