+++ rte/_buildrun.sh
@@ -2,5 +2,6 @@ git submodule update --init --recursive
 mkdir build
 cd build
 cmake .. -DGENERATE_WRAPPER=ON
-make
-./brux
+cd ..
+cmake --build build
+./build/brux
